// src/models/create-fixture-dto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * GoMoney Basic Football-League Manager Service
 * This service provides endpoints for all `basic football-league manager` related interactions
 *
 * OpenAPI spec version: 1.0.0
 * Contact: fatai@mail.com
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
/**
 * Create Fixture request data
 * @export
 * @interface CreateFixtureDto
 */
export interface CreateFixtureDto {
    /**
     * ID of Home Team
     * @type {string}
     * @memberof CreateFixtureDto
     */
    homeTeamId: string;
    /**
     * ID of Away Team
     * @type {string}
     * @memberof CreateFixtureDto
     */
    awayTeamId: string;
    /**
     * Date & Time that match starts
     * @type {Date}
     * @memberof CreateFixtureDto
     */
    commencesAt: Date;
}